chore: split monolith into separate files
