+++ pkg/provider/vault/util/vault.go
@@ -50,7 +50,7 @@ type Client interface {
 	AddHeader(key, value string)
 }
 
-type VClient struct {
+type VaultClient struct {
 	SetTokenFunc     func(v string)
 	TokenFunc        func() string
 	ClearTokenFunc   func()
@@ -61,34 +61,34 @@ type VClient struct {
 	AddHeaderFunc    func(key, value string)
 }
 
-func (v VClient) AddHeader(key, value string) {
+func (v VaultClient) AddHeader(key, value string) {
 	v.AddHeaderFunc(key, value)
 }
 
-func (v VClient) SetNamespace(namespace string) {
+func (v VaultClient) SetNamespace(namespace string) {
 	v.SetNamespaceFunc(namespace)
 }
 
-func (v VClient) ClearToken() {
+func (v VaultClient) ClearToken() {
 	v.ClearTokenFunc()
 }
 
-func (v VClient) Token() string {
+func (v VaultClient) Token() string {
 	return v.TokenFunc()
 }
 
-func (v VClient) SetToken(token string) {
+func (v VaultClient) SetToken(token string) {
 	v.SetTokenFunc(token)
 }
 
-func (v VClient) Auth() Auth {
+func (v VaultClient) Auth() Auth {
 	return v.AuthField
 }
 
-func (v VClient) AuthToken() Token {
+func (v VaultClient) AuthToken() Token {
 	return v.AuthTokenField
 }
 
-func (v VClient) Logical() Logical {
+func (v VaultClient) Logical() Logical {
 	return v.LogicalField
 }
